Update landing page to explain "open-source" delay
Seems that source code is on a delay, meaning the source in the various app stores isn't the same as the source in GitHub. As a result, quite a few people have been concerned (e.g. https://github.com/Coinomi/coinomi-android/issues/204). This clarifies the stance on the landing page. Feel free to reject, but wanted to at least put this out there.